+++ docs/griptape-framework/drivers/src/prompt_drivers_structured_output.py
@@ -1,0 +1,35 @@
+import schema
+from rich.pretty import pprint
+
+from griptape.drivers import OpenAiChatPromptDriver
+from griptape.rules import JsonSchemaRule, Rule
+from griptape.structures import Pipeline
+from griptape.tasks import PromptTask
+
+pipeline = Pipeline(
+    tasks=[
+        PromptTask(
+            prompt_driver=OpenAiChatPromptDriver(
+                model="gpt-4o",
+                use_native_structured_output=True,
+                native_structured_output_strategy="native",
+            ),
+            rules=[
+                Rule("You are a helpful math tutor. Guide the user through the solution step by step."),
+                JsonSchemaRule(
+                    schema.Schema(
+                        {
+                            "steps": [schema.Schema({"explanation": str, "output": str})],
+                            "final_answer": str,
+                        }
+                    )
+                ),
+            ],
+        )
+    ]
+)
+
+output = pipeline.run("How can I solve 8x + 7 = -23").output.value
+
+
+pprint(output)