+++ tests/unit/configs/drivers/test_amazon_bedrock_drivers_config.py
@@ -51,6 +51,8 @@ class TestAmazonBedrockDriversConfig:
                 "type": "AmazonBedrockPromptDriver",
                 "tool_choice": {"auto": {}},
                 "use_native_tools": True,
+                "use_native_structured_output": True,
+                "native_structured_output_strategy": "tool",
                 "extra_params": {},
             },
             "vector_store_driver": {
@@ -106,6 +108,8 @@ class TestAmazonBedrockDriversConfig:
                 "type": "AmazonBedrockPromptDriver",
                 "tool_choice": {"auto": {}},
                 "use_native_tools": True,
+                "use_native_structured_output": True,
+                "native_structured_output_strategy": "tool",
                 "extra_params": {},
             },
             "vector_store_driver": {