+++ tests/unit/configs/drivers/test_anthropic_drivers_config.py
@@ -25,6 +25,8 @@ class TestAnthropicDriversConfig:
                 "top_p": 0.999,
                 "top_k": 250,
                 "use_native_tools": True,
+                "native_structured_output_strategy": "tool",
+                "use_native_structured_output": True,
                 "extra_params": {},
             },
             "image_generation_driver": {"type": "DummyImageGenerationDriver"},