+++ tests/unit/drivers/prompt/test_cohere_prompt_driver.py
@@ -2,6 +2,7 @@ import json
 from unittest.mock import Mock
 
 import pytest
+from schema import Schema
 
 from griptape.artifacts.action_artifact import ActionArtifact
 from griptape.artifacts.list_artifact import ListArtifact
@@ -12,6 +13,36 @@ from tests.mocks.mock_tool.tool import MockTool
 
 
 class TestCoherePromptDriver:
+    COHERE_STRUCTURED_OUTPUT_SCHEMA = {
+        "$id": "Output",
+        "$schema": "http://json-schema.org/draft-07/schema#",
+        "additionalProperties": False,
+        "properties": {"foo": {"type": "string"}},
+        "required": ["foo"],
+        "type": "object",
+    }
+    COHERE_STRUCTURED_OUTPUT_TOOL = {
+        "function": {
+            "description": "Used to provide the final response which ends this conversation.",
+            "name": "StructuredOutputTool_provide_output",
+            "parameters": {
+                "$id": "Parameters Schema",
+                "$schema": "http://json-schema.org/draft-07/schema#",
+                "additionalProperties": False,
+                "properties": {
+                    "values": {
+                        "additionalProperties": False,
+                        "properties": {"foo": {"type": "string"}},
+                        "required": ["foo"],
+                        "type": "object",
+                    },
+                },
+                "required": ["values"],
+                "type": "object",
+            },
+        },
+        "type": "function",
+    }
     COHERE_TOOLS = [
         {
             "function": {
@@ -242,6 +273,7 @@ class TestCoherePromptDriver:
     @pytest.fixture()
     def prompt_stack(self):
         prompt_stack = PromptStack()
+        prompt_stack.output_schema = Schema({"foo": str})
         prompt_stack.tools = [MockTool()]
         prompt_stack.add_system_message("system-input")
         prompt_stack.add_user_message("user-input")
@@ -306,10 +338,25 @@ class TestCoherePromptDriver:
         assert CoherePromptDriver(model="command", api_key="foobar")
 
     @pytest.mark.parametrize("use_native_tools", [True, False])
-    def test_try_run(self, mock_client, prompt_stack, messages, use_native_tools):
+    @pytest.mark.parametrize("use_native_structured_output", [True, False])
+    @pytest.mark.parametrize("native_structured_output_strategy", ["native", "tool", "foo"])
+    def test_try_run(
+        self,
+        mock_client,
+        prompt_stack,
+        messages,
+        use_native_tools,
+        use_native_structured_output,
+        native_structured_output_strategy,
+    ):
         # Given
         driver = CoherePromptDriver(
-            model="command", api_key="api-key", use_native_tools=use_native_tools, extra_params={"foo": "bar"}
+            model="command",
+            api_key="api-key",
+            use_native_tools=use_native_tools,
+            use_native_structured_output=use_native_structured_output,
+            native_structured_output_strategy=native_structured_output_strategy,
+            extra_params={"foo": "bar"},
         )
 
         # When
@@ -320,7 +367,26 @@ class TestCoherePromptDriver:
             model="command",
             messages=messages,
             max_tokens=None,
-            **({"tools": self.COHERE_TOOLS} if use_native_tools else {}),
+            **{
+                "tools": [
+                    *self.COHERE_TOOLS,
+                    *(
+                        [self.COHERE_STRUCTURED_OUTPUT_TOOL]
+                        if use_native_structured_output and native_structured_output_strategy == "tool"
+                        else []
+                    ),
+                ]
+            }
+            if use_native_tools
+            else {},
+            **{
+                "response_format": {
+                    "type": "json_object",
+                    "schema": self.COHERE_STRUCTURED_OUTPUT_SCHEMA,
+                }
+            }
+            if use_native_structured_output and native_structured_output_strategy == "native"
+            else {},
             stop_sequences=[],
             temperature=0.1,
             foo="bar",
@@ -340,13 +406,25 @@ class TestCoherePromptDriver:
         assert message.usage.output_tokens == 10
 
     @pytest.mark.parametrize("use_native_tools", [True, False])
-    def test_try_stream_run(self, mock_stream_client, prompt_stack, messages, use_native_tools):
+    @pytest.mark.parametrize("use_native_structured_output", [True, False])
+    @pytest.mark.parametrize("native_structured_output_strategy", ["native", "tool", "foo"])
+    def test_try_stream_run(
+        self,
+        mock_stream_client,
+        prompt_stack,
+        messages,
+        use_native_tools,
+        use_native_structured_output,
+        native_structured_output_strategy,
+    ):
         # Given
         driver = CoherePromptDriver(
             model="command",
             api_key="api-key",
             stream=True,
             use_native_tools=use_native_tools,
+            use_native_structured_output=use_native_structured_output,
+            native_structured_output_strategy=native_structured_output_strategy,
             extra_params={"foo": "bar"},
         )
 
@@ -359,7 +437,26 @@ class TestCoherePromptDriver:
             model="command",
             messages=messages,
             max_tokens=None,
-            **({"tools": self.COHERE_TOOLS} if use_native_tools else {}),
+            **{
+                "tools": [
+                    *self.COHERE_TOOLS,
+                    *(
+                        [self.COHERE_STRUCTURED_OUTPUT_TOOL]
+                        if use_native_structured_output and native_structured_output_strategy == "tool"
+                        else []
+                    ),
+                ]
+            }
+            if use_native_tools
+            else {},
+            **{
+                "response_format": {
+                    "type": "json_object",
+                    "schema": self.COHERE_STRUCTURED_OUTPUT_SCHEMA,
+                }
+            }
+            if use_native_structured_output and native_structured_output_strategy == "native"
+            else {},
             stop_sequences=[],
             temperature=0.1,
             foo="bar",