+++ tests/unit/drivers/prompt/test_google_prompt_driver.py
@@ -4,6 +4,7 @@ import pytest
 from google.generativeai.protos import FunctionCall, FunctionResponse, Part
 from google.generativeai.types import ContentDict, GenerationConfig
 from google.protobuf.json_format import MessageToDict
+from schema import Schema
 
 from griptape.artifacts import ActionArtifact, GenericArtifact, ImageArtifact, TextArtifact
 from griptape.artifacts.list_artifact import ListArtifact
@@ -13,6 +14,15 @@ from tests.mocks.mock_tool.tool import MockTool
 
 
 class TestGooglePromptDriver:
+    GOOGLE_STRUCTURED_OUTPUT_TOOL = {
+        "description": "Used to provide the final response which ends this conversation.",
+        "name": "StructuredOutputTool_provide_output",
+        "parameters": {
+            "properties": {"foo": {"type": "STRING"}},
+            "required": ["foo"],
+            "type": "OBJECT",
+        },
+    }
     GOOGLE_TOOLS = [
         {
             "name": "MockTool_test",
@@ -100,6 +110,7 @@ class TestGooglePromptDriver:
     @pytest.fixture(params=[True, False])
     def prompt_stack(self, request):
         prompt_stack = PromptStack()
+        prompt_stack.output_schema = Schema({"foo": str})
         prompt_stack.tools = [MockTool()]
         if request.param:
             prompt_stack.add_system_message("system-input")
@@ -166,7 +177,10 @@ class TestGooglePromptDriver:
         assert driver
 
     @pytest.mark.parametrize("use_native_tools", [True, False])
-    def test_try_run(self, mock_generative_model, prompt_stack, messages, use_native_tools):
+    @pytest.mark.parametrize("use_native_structured_output", [True, False])
+    def test_try_run(
+        self, mock_generative_model, prompt_stack, messages, use_native_tools, use_native_structured_output
+    ):
         # Given
         driver = GooglePromptDriver(
             model="gemini-pro",
@@ -174,6 +188,8 @@ class TestGooglePromptDriver:
             top_p=0.5,
             top_k=50,
             use_native_tools=use_native_tools,
+            use_native_structured_output=use_native_structured_output,
+            native_structured_output_strategy="tool",
             extra_params={"max_output_tokens": 10},
         )
 
@@ -195,9 +211,14 @@ class TestGooglePromptDriver:
         )
         if use_native_tools:
             tool_declarations = call_args.kwargs["tools"]
-            assert [
-                MessageToDict(tool_declaration.to_proto()._pb) for tool_declaration in tool_declarations
-            ] == self.GOOGLE_TOOLS
+            tools = [
+                *self.GOOGLE_TOOLS,
+                *([self.GOOGLE_STRUCTURED_OUTPUT_TOOL] if use_native_structured_output else []),
+            ]
+            assert [MessageToDict(tool_declaration.to_proto()._pb) for tool_declaration in tool_declarations] == tools
+
+            if use_native_structured_output:
+                assert call_args.kwargs["tool_config"] == {"function_calling_config": {"mode": "auto"}}
 
         assert isinstance(message.value[0], TextArtifact)
         assert message.value[0].value == "model-output"
@@ -210,7 +231,10 @@ class TestGooglePromptDriver:
         assert message.usage.output_tokens == 10
 
     @pytest.mark.parametrize("use_native_tools", [True, False])
-    def test_try_stream(self, mock_stream_generative_model, prompt_stack, messages, use_native_tools):
+    @pytest.mark.parametrize("use_native_structured_output", [True, False])
+    def test_try_stream(
+        self, mock_stream_generative_model, prompt_stack, messages, use_native_tools, use_native_structured_output
+    ):
         # Given
         driver = GooglePromptDriver(
             model="gemini-pro",
@@ -219,6 +243,7 @@ class TestGooglePromptDriver:
             top_p=0.5,
             top_k=50,
             use_native_tools=use_native_tools,
+            use_native_structured_output=use_native_structured_output,
             extra_params={"max_output_tokens": 10},
         )
 
@@ -242,9 +267,14 @@ class TestGooglePromptDriver:
         )
         if use_native_tools:
             tool_declarations = call_args.kwargs["tools"]
-            assert [
-                MessageToDict(tool_declaration.to_proto()._pb) for tool_declaration in tool_declarations
-            ] == self.GOOGLE_TOOLS
+            tools = [
+                *self.GOOGLE_TOOLS,
+                *([self.GOOGLE_STRUCTURED_OUTPUT_TOOL] if use_native_structured_output else []),
+            ]
+            assert [MessageToDict(tool_declaration.to_proto()._pb) for tool_declaration in tool_declarations] == tools
+
+            if use_native_structured_output:
+                assert call_args.kwargs["tool_config"] == {"function_calling_config": {"mode": "auto"}}
         assert isinstance(event.content, TextDeltaMessageContent)
         assert event.content.text == "model-output"
         assert event.usage.input_tokens == 5
@@ -259,3 +289,9 @@ class TestGooglePromptDriver:
 
         event = next(stream)
         assert event.usage.output_tokens == 5
+
+    def test_verify_native_structured_output_strategy(self):
+        assert GooglePromptDriver(model="foo", native_structured_output_strategy="tool")
+
+        with pytest.raises(ValueError, match="GooglePromptDriver does not support `native` structured output mode."):
+            GooglePromptDriver(model="foo", native_structured_output_strategy="native")