+++ tests/unit/drivers/prompt/test_openai_chat_prompt_driver.py
@@ -12,6 +12,36 @@ from tests.mocks.mock_tool.tool import MockTool
 
 
 class TestOpenAiChatPromptDriverFixtureMixin:
+    OPENAI_STRUCTURED_OUTPUT_SCHEMA = {
+        "$id": "Output",
+        "$schema": "http://json-schema.org/draft-07/schema#",
+        "additionalProperties": False,
+        "properties": {"foo": {"type": "string"}},
+        "required": ["foo"],
+        "type": "object",
+    }
+    OPENAI_STRUCTURED_OUTPUT_TOOL = {
+        "function": {
+            "description": "Used to provide the final response which ends this conversation.",
+            "name": "StructuredOutputTool_provide_output",
+            "parameters": {
+                "$id": "Parameters Schema",
+                "$schema": "http://json-schema.org/draft-07/schema#",
+                "additionalProperties": False,
+                "properties": {
+                    "values": {
+                        "additionalProperties": False,
+                        "properties": {"foo": {"type": "string"}},
+                        "required": ["foo"],
+                        "type": "object",
+                    },
+                },
+                "required": ["values"],
+                "type": "object",
+            },
+        },
+        "type": "function",
+    }
     OPENAI_TOOLS = [
         {
             "function": {
@@ -239,6 +269,7 @@ class TestOpenAiChatPromptDriverFixtureMixin:
     @pytest.fixture()
     def prompt_stack(self):
         prompt_stack = PromptStack()
+        prompt_stack.output_schema = schema.Schema({"foo": str})
         prompt_stack.tools = [MockTool()]
         prompt_stack.add_system_message("system-input")
         prompt_stack.add_user_message("user-input")
@@ -340,11 +371,23 @@ class TestOpenAiChatPromptDriver(TestOpenAiChatPromptDriverFixtureMixin):
         assert OpenAiChatPromptDriver(model=OpenAiTokenizer.DEFAULT_OPENAI_GPT_4_MODEL)
 
     @pytest.mark.parametrize("use_native_tools", [True, False])
-    def test_try_run(self, mock_chat_completion_create, prompt_stack, messages, use_native_tools):
+    @pytest.mark.parametrize("use_native_structured_output", [True, False])
+    @pytest.mark.parametrize("native_structured_output_strategy", ["native", "tool", "foo"])
+    def test_try_run(
+        self,
+        mock_chat_completion_create,
+        prompt_stack,
+        messages,
+        use_native_tools,
+        use_native_structured_output,
+        native_structured_output_strategy,
+    ):
         # Given
         driver = OpenAiChatPromptDriver(
             model=OpenAiTokenizer.DEFAULT_OPENAI_GPT_3_CHAT_MODEL,
             use_native_tools=use_native_tools,
+            use_native_structured_output=use_native_structured_output,
+            native_structured_output_strategy=native_structured_output_strategy,
             extra_params={"foo": "bar"},
         )
 
@@ -359,11 +402,32 @@ class TestOpenAiChatPromptDriver(TestOpenAiChatPromptDriverFixtureMixin):
             messages=messages,
             seed=driver.seed,
             **{
-                "tools": self.OPENAI_TOOLS,
-                "tool_choice": driver.tool_choice,
+                "tools": [
+                    *self.OPENAI_TOOLS,
+                    *(
+                        [self.OPENAI_STRUCTURED_OUTPUT_TOOL]
+                        if use_native_structured_output and native_structured_output_strategy == "tool"
+                        else []
+                    ),
+                ],
+                "tool_choice": "required"
+                if use_native_structured_output and native_structured_output_strategy == "tool"
+                else driver.tool_choice,
                 "parallel_tool_calls": driver.parallel_tool_calls,
             }
             if use_native_tools
+            else {},
+            **{
+                "response_format": {
+                    "type": "json_schema",
+                    "json_schema": {
+                        "name": "Output",
+                        "schema": self.OPENAI_STRUCTURED_OUTPUT_SCHEMA,
+                        "strict": True,
+                    },
+                }
+            }
+            if use_native_structured_output and native_structured_output_strategy == "native"
             else {},
             foo="bar",
         )
@@ -445,12 +509,24 @@ class TestOpenAiChatPromptDriver(TestOpenAiChatPromptDriverFixtureMixin):
         assert message.usage.output_tokens == 10
 
     @pytest.mark.parametrize("use_native_tools", [True, False])
-    def test_try_stream_run(self, mock_chat_completion_stream_create, prompt_stack, messages, use_native_tools):
+    @pytest.mark.parametrize("use_native_structured_output", [True, False])
+    @pytest.mark.parametrize("native_structured_output_strategy", ["native", "tool", "foo"])
+    def test_try_stream_run(
+        self,
+        mock_chat_completion_stream_create,
+        prompt_stack,
+        messages,
+        use_native_tools,
+        use_native_structured_output,
+        native_structured_output_strategy,
+    ):
         # Given
         driver = OpenAiChatPromptDriver(
             model=OpenAiTokenizer.DEFAULT_OPENAI_GPT_3_CHAT_MODEL,
             stream=True,
             use_native_tools=use_native_tools,
+            use_native_structured_output=use_native_structured_output,
+            native_structured_output_strategy=native_structured_output_strategy,
             extra_params={"foo": "bar"},
         )
 
@@ -468,11 +544,32 @@ class TestOpenAiChatPromptDriver(TestOpenAiChatPromptDriverFixtureMixin):
             seed=driver.seed,
             stream_options={"include_usage": True},
             **{
-                "tools": self.OPENAI_TOOLS,
-                "tool_choice": driver.tool_choice,
+                "tools": [
+                    *self.OPENAI_TOOLS,
+                    *(
+                        [self.OPENAI_STRUCTURED_OUTPUT_TOOL]
+                        if use_native_structured_output and native_structured_output_strategy == "tool"
+                        else []
+                    ),
+                ],
+                "tool_choice": "required"
+                if use_native_structured_output and native_structured_output_strategy == "tool"
+                else driver.tool_choice,
                 "parallel_tool_calls": driver.parallel_tool_calls,
             }
             if use_native_tools
+            else {},
+            **{
+                "response_format": {
+                    "type": "json_schema",
+                    "json_schema": {
+                        "name": "Output",
+                        "schema": self.OPENAI_STRUCTURED_OUTPUT_SCHEMA,
+                        "strict": True,
+                    },
+                }
+            }
+            if use_native_structured_output and native_structured_output_strategy == "native"
             else {},
             foo="bar",
         )
@@ -500,7 +597,10 @@ class TestOpenAiChatPromptDriver(TestOpenAiChatPromptDriverFixtureMixin):
     def test_try_run_with_max_tokens(self, mock_chat_completion_create, prompt_stack, messages):
         # Given
         driver = OpenAiChatPromptDriver(
-            model=OpenAiTokenizer.DEFAULT_OPENAI_GPT_3_CHAT_MODEL, max_tokens=1, use_native_tools=False
+            model=OpenAiTokenizer.DEFAULT_OPENAI_GPT_3_CHAT_MODEL,
+            max_tokens=1,
+            use_native_tools=False,
+            use_native_structured_output=False,
         )
 
         # When
@@ -535,6 +635,7 @@ class TestOpenAiChatPromptDriver(TestOpenAiChatPromptDriverFixtureMixin):
             tokenizer=MockTokenizer(model="mock-model", stop_sequences=["mock-stop"]),
             max_tokens=1,
             use_native_tools=False,
+            use_native_structured_output=False,
         )
 
         # When