+++ tests/unit/tasks/test_prompt_task.py
@@ -1,9 +1,15 @@
+import warnings
+
+import pytest
+
 from griptape.artifacts.image_artifact import ImageArtifact
+from griptape.artifacts.json_artifact import JsonArtifact
 from griptape.artifacts.list_artifact import ListArtifact
 from griptape.artifacts.text_artifact import TextArtifact
 from griptape.memory.structure import ConversationMemory
 from griptape.memory.structure.run import Run
 from griptape.rules import Rule
+from griptape.rules.json_schema_rule import JsonSchemaRule
 from griptape.rules.ruleset import Ruleset
 from griptape.structures import Pipeline
 from griptape.tasks import PromptTask
@@ -171,6 +177,81 @@ class TestPromptTask:
         assert task.prompt_stack.messages[1].to_text() == "output"
         assert task.prompt_stack.messages[2].is_user()
         assert task.prompt_stack.messages[2].to_text() == "test value"
+
+    def test_prompt_stack_native_schema(self):
+        from schema import Schema
+
+        output_schema = Schema({"baz": str})
+        task = PromptTask(
+            input="foo",
+            prompt_driver=MockPromptDriver(
+                use_native_structured_output=True,
+                mock_structured_output={"baz": "foo"},
+            ),
+            rules=[JsonSchemaRule(output_schema)],
+        )
+        output = task.run()
+
+        assert isinstance(output, JsonArtifact)
+        assert output.value == {"baz": "foo"}
+
+        assert task.prompt_stack.output_schema is output_schema
+        assert task.prompt_stack.messages[0].is_user()
+        assert "foo" in task.prompt_stack.messages[0].to_text()
+
+        # Ensure no warnings were raised
+        with warnings.catch_warnings():
+            warnings.simplefilter("error")
+            assert task.prompt_stack
+
+    def test_prompt_stack_mixed_native_schema(self):
+        from schema import Schema
+
+        output_schema = Schema({"baz": str})
+        task = PromptTask(
+            input="foo",
+            prompt_driver=MockPromptDriver(
+                use_native_structured_output=True,
+            ),
+            rules=[Rule("foo"), JsonSchemaRule({"bar": {}}), JsonSchemaRule(output_schema)],
+        )
+
+        assert task.prompt_stack.output_schema is output_schema
+        assert task.prompt_stack.messages[0].is_system()
+        assert "foo" in task.prompt_stack.messages[0].to_text()
+        assert "bar" not in task.prompt_stack.messages[0].to_text()
+        with pytest.warns(
+            match="Not all provided `JsonSchemaRule`s include a `schema.Schema` instance. These will be ignored with `use_native_structured_output`."
+        ):
+            assert task.prompt_stack
+
+    def test_prompt_stack_empty_native_schema(self):
+        task = PromptTask(
+            input="foo",
+            prompt_driver=MockPromptDriver(
+                use_native_structured_output=True,
+            ),
+            rules=[JsonSchemaRule({"foo": {}})],
+        )
+
+        assert task.prompt_stack.output_schema is None
+
+    def test_prompt_stack_multi_native_schema(self):
+        from schema import Or, Schema
+
+        output_schema = Schema({"foo": str})
+        task = PromptTask(
+            input="foo",
+            prompt_driver=MockPromptDriver(
+                use_native_structured_output=True,
+            ),
+            rules=[JsonSchemaRule({"foo": {}}), JsonSchemaRule(output_schema), JsonSchemaRule(output_schema)],
+        )
+
+        assert isinstance(task.prompt_stack.output_schema, Schema)
+        assert task.prompt_stack.output_schema.json_schema("Output") == Schema(
+            Or(output_schema, output_schema)
+        ).json_schema("Output")
 
     def test_rulesets(self):
         pipeline = Pipeline(