+++ CHANGELOG.md
@@ -31,6 +31,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Structure.run_stream()` for streaming Events from a Structure as an iterator.
 - Support for `GenericMessageContent` in `AnthropicPromptDriver` and `AmazonBedrockPromptDriver`.
 - Validators to `Agent` initialization.
+- `BasePromptDriver.use_native_structured_output` for enabling or disabling structured output.
+- `BasePromptDriver.native_structured_output_strategy` for changing the structured output strategy between `native` and `tool`.
 
 ### Changed
 
@@ -39,6 +41,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `PromptTask.prompt_driver` is now serialized.
 - `PromptTask` can now do everything a `ToolkitTask` can do.
 - Loosten `numpy`s version constraint to `>=1.26.4,<3`.
+- `JsonSchemaRule`s can now take a `schema.Schema` instance. Required for using a `JsonSchemaRule` with structured output.
+- `JsonSchemaRule`s will now be used for structured output if the Prompt Driver supports it.
 
 ### Fixed